services/bidding-server/pyproject.toml
[tool.poetry]
name = "bidding-server"
version = "0.1.0"
description = ""
authors = ["ray <wuhaohsiang1992@gmail.com>"]
readme = "README.md"
packages = [{include = "bidding_server"}]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
requests = "^2.28.2"
fastapi = "^0.93.0"
psycopg2-binary = "^2.9.5"
rich = "^13.3.2"
uvicorn = "^0.20.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
